+++ src-core/src/discord/commands.rs
@@ -6,6 +6,12 @@ pub async fn discord_update_activity(
     asset: String,
     asset_label: Option<String>,
 ) -> bool {
-    let result = super::update_activity(details, state, asset, asset_label).await;
-    result
+    super::update_activity(details, state, asset, asset_label).await
+}
+
+#[tauri::command]
+#[oyasumivr_macros::command_profiling]
+pub async fn discord_clear_activity(
+) -> bool {
+    super::clear_activity().await
 }