+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.12.3]
+
+### Added
+
+- Options for (conditionally) disabling Discord rich presence, based on sleep mode, or VRChat being active.
+- Missing event logs for Bigscreen Beyond fan- and led automations.
+- Missing Dutch translations
+
+### Fixed
+
+- Fixed sleep preparation not being available even when some of its automations are enabled.
+
 ## [1.12.2]
 
 ### Changed